+++ src/types.ts
@@ -7,9 +7,11 @@ export interface CalendarHeatmapOptions {
   // Data Links
   dataLinks?: CalendarDataLink[];
   // Colors
-  colorScheme: 'green' | 'blue' | 'red' | 'yellow' | 'purple' | 'orange' | 'custom';
+  colorScheme: 'green' | 'blue' | 'red' | 'yellow' | 'purple' | 'orange' | 'custom' | 'custom-gradient';
   emptyColor: string;
   customColor: string;
+  gradientColorLow: string;
+  gradientColorHigh: string;
 
   // Layout
   autoRectSize: boolean;